requirements.txt
Flask==3.0.1
opencv-python==4.9.0.80
mediapipe==0.10.9
Pillow==10.2.0
numpy==1.26.3
